staticfiles/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-unfold==0.89.0
djangorestframework==3.17.1
gunicorn==25.3.0
packaging==26.0
pillow==12.2.0
PyMuPDF==1.27.2.2
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
